requirements.txt
python-dotenv
requests
pydantic
openai
flask
